pyproject.toml
# pyproject.toml
#
# This file defines the build configuration for the RAG industrial project.
# It follows the PEP 621 specification and uses Hatchling to build the
# package.  Dependencies are split into runtime and development groups.

[build-system]
requires = [
  "hatchling>=1.17.0"
]
build-backend = "hatchling.build"

[project]
name = "rag-industrial-eval-guardrails"
version = "0.1.0"
description = "An industrial-grade RAG pipeline with evaluation and guardrails."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

# Runtime dependencies required to run the service.  Where possible
# lightweight alternatives are used to keep the container small.  The
# version constraints reflect combinations known to work together at
# the time of writing.  Updating versions may require adjusting API
# usage.
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "asyncpg>=0.29.0",
  "sqlalchemy>=2.0.29",
  "psycopg[binary]>=3.1.18",
  "pgvector>=0.2.4",
  "openai>=1.13.3",
  "sentence-transformers>=2.3.1",
  "transformers>=4.39.3",
  "ragas>=0.1.3",
  "datasets>=2.18.0",
  "presidio-analyzer>=2.2.359",
  "presidio-anonymizer>=2.2.359",
  "spacy>=3.7.2",
  "tiktoken>=0.6.0",
  "pydantic>=2.7.1",
  "python-dotenv>=1.0.1",
  "structlog>=23.2.0",
  "prometheus-client>=0.19.0",
  "tenacity>=8.2.2",
  "aiohttp>=3.9.4",
  "rich>=13.7.0"
  ,
  # Utilities for generating and parsing PDF/HTML documents in the synthetic corpus
  "fpdf2>=2.7.7",
  "pdfminer.six>=20221105",
  "beautifulsoup4>=4.12.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.6",
  "requests>=2.31.0",
  "black>=24.3.0",
  "ruff>=0.3.7",
  "mypy>=1.8.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "types-requests>=2.31.0.20240404"
]

[tool.hatch.metadata]
allow-direct-references = true
